add skip-promt for destroying node

+++ down.go
@@ -4,12 +4,31 @@ import (
 	"github.com/urfave/cli"
 	"log"
 	"os/exec"
+	"bufio"
+	"os"
+	"fmt"
+	"strings"
 )
 
 // destroyNode will tear down the deployed darknode, but keep the config file.
 func destroyNode(ctx *cli.Context) error {
 	// FIXME : currently it only supports tear down AWS deployment.
 	// Needs to figure out way which suits for all kinds of cloud service.
+	skip := ctx.Bool("skip-prompt")
+	if !skip {
+		for {
+			fmt.Println("Have you deregistered your node and withdrawn all fees? (Yes/No)")
+			reader := bufio.NewReader(os.Stdin)
+			text, _ := reader.ReadString('\n')
+			if strings.ToLower(strings.TrimSpace(text)) == "yes"{
+				break
+			}
+			if strings.ToLower(strings.TrimSpace(text)) == "no"{
+				return nil
+			}
+		}
+	}
+
 	return destroyAwsNode()
 }
